requirements.txt
PyYAML==6.0.1
click==8.1.7
feedparser==6.0.11
APScheduler==3.10.4
aiofile==3.8.8
httpx[http2]==0.27.2
pydantic==2.9.2